requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.2
requests>=2.28
python-dotenv>=1.0
hmmlearn>=0.3
# Dashboard dependencies
streamlit>=1.28
plotly>=5.18
